Collapse GuardedInvoker into Broker and remove MCP direct handler
Replace GuardedInvoker with a simpler AuditedInvoker that handles only depth/recursion checks and audit logging. The rate limiter, provider allowlist, and options machinery are deleted entirely since WithoutRateLimit() was always used in production, and neither WithMaxDepth nor WithAllowedProviders were used outside tests. The new AuditedInvoker wraps a Broker with a source string and audit sink, reducing the abstraction to its essential behavior.

Remove the makeDirectHandler code path from the MCP layer. Previously, providers implementing directToolCaller were routed through a separate handler that bypassed the broker and manually duplicated audit logging. Now all MCP tool calls go through the unified makeHandler -> invoker -> Broker path, with the broker's existing TransportMCPPassthrough support handling direct tool calls via CallDirectTool. MCP request metadata is propagated through context so the broker can forward it to upstream providers.
