+++ gestaltd/internal/mcp/binding.go
@@ -17,10 +17,6 @@ const (
 	toolNameSep   = "_"
 )
 
-type directToolCaller interface {
-	CallTool(ctx context.Context, name string, args map[string]any) (*mcpgo.CallToolResult, error)
-}
-
 type Config struct {
 	Invoker          invocation.Invoker
 	TokenResolver    invocation.TokenResolver
@@ -65,7 +61,7 @@ func NewServer(cfg Config) *mcpserver.MCPServer {
 		}
 
 		if cat := prov.Catalog(); cat != nil {
-			addCatalogTools(srv, cfg, provName, cat, prov)
+			addCatalogTools(srv, cfg, provName, cat)
 		}
 	}
